Scrapped Images/model/code/requirements.txt.txt
fastai
torchvision
pillow